+++ code/packages/store-azure-web-pubsub/src/index.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-export { WeaveStoreAzureWebPubsub } from "./store-azure-web-pubsub";
-export { WeaveStoreAzureWebPubSubSyncHost } from "./host";
-export * from "./constants";
-export * from "./types";
+export { WeaveStoreAzureWebPubsub } from './store-azure-web-pubsub';
+export { WeaveAzureWebPubsubServer } from './server/azure-web-pubsub-server';
+export * from './constants';
+export * from './types';