+++ code/CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.0] - 2025-04-14
+
+### Added
+
+- [#95](https://github.com/InditexTech/weavejs/issues/95) Refactor server part for websockets store
+- [#97](https://github.com/InditexTech/weavejs/issues/97) Refactor server part for Azure web pubsub store
+
 ## [0.23.0] - 2025-04-10
 
 ### Added
@@ -199,7 +206,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#8](https://github.com/InditexTech/weavejs/issues/8) ESLint OSS rules
 - [#2](https://github.com/InditexTech/weavejs/issues/2) Improve rendering handling
 
-[Unreleased]: https://github.com/InditexTech/weavejs/compare/0.23.0...HEAD
+[Unreleased]: https://github.com/InditexTech/weavejs/compare/0.24.0...HEAD
+
+[0.24.0]: https://github.com/InditexTech/weavejs/compare/0.23.0...0.24.0
 
 [0.23.0]: https://github.com/InditexTech/weavejs/compare/0.22.0...0.23.0